# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Haraj Station API
  version: 1.0.0
  description: Generated inventory of route handlers. Request and response schemas are being added per mobile contract.
servers:
  - url: https://harajstation.com
paths:
...
  "/api/listings":
    post:
      summary: "POST /api/listings"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/mobile/auctions":
    get:
      summary: "GET /api/mobile/auctions"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...
  "/api/mobile/listings/{id}":
    get:
      summary: "GET /api/mobile/listings/{id}"
      responses:
        "200":
          description: Successful response
        "4XX":
          description: Error response
...